A major bug is resolved w.r.t precision loss of Tags in InstaTag when tags needs to be shown after tagging photo.
Solved with the use of float in the Tag Model and Instead of Setting out Layout Params while adding tag setX and setY.

With this bug resolution a major bug has been resolved in the app which is heading towards creation of other interesting repository for instagram clone after this InstaTag Library.

+++ instatag/src/main/java/com/harsh/instatag/Tag.java
@@ -21,10 +21,10 @@ import android.os.Parcelable;
 
 public class Tag implements Parcelable {
     private String unique_tag_id;
-    private Double x_co_ord;
-    private Double y_co_ord;
+    private float x_co_ord;
+    private float y_co_ord;
 
-    public Tag(String unique_tag_id, Double x_co_ord, Double y_co_ord) {
+    public Tag(String unique_tag_id, float x_co_ord, float y_co_ord) {
         this.unique_tag_id = unique_tag_id;
         this.x_co_ord = x_co_ord;
         this.y_co_ord = y_co_ord;
@@ -39,19 +39,19 @@ public class Tag implements Parcelable {
         this.unique_tag_id = unique_tag_id;
     }
 
-    public Double getX_co_ord() {
+    public float getX_co_ord() {
         return x_co_ord;
     }
 
-    public void setX_co_ord(Double x_co_ord) {
+    public void setX_co_ord(float x_co_ord) {
         this.x_co_ord = x_co_ord;
     }
 
-    public Double getY_co_ord() {
+    public float getY_co_ord() {
         return y_co_ord;
     }
 
-    public void setY_co_ord(Double y_co_ord) {
+    public void setY_co_ord(float y_co_ord) {
         this.y_co_ord = y_co_ord;
     }
 
@@ -70,8 +70,8 @@ public class Tag implements Parcelable {
 
     private Tag(Parcel in) {
         this.unique_tag_id = in.readString();
-        this.x_co_ord = (Double) in.readValue(Double.class.getClassLoader());
-        this.y_co_ord = (Double) in.readValue(Double.class.getClassLoader());
+        this.x_co_ord = (float) in.readValue(float.class.getClassLoader());
+        this.y_co_ord = (float) in.readValue(float.class.getClassLoader());
     }
 
     public static final Parcelable.Creator<Tag> CREATOR =
